refactor: reuse cache config on chunk cache
Extend cache config with chunk cache configs (default and disk-based config) to reuse the configuration definitions.
